Guard resolve with runtime feature.
