+++ compio-net/src/resolve/mod.rs
@@ -11,7 +11,13 @@ cfg_if::cfg_if! {
     }
 }
 
-use std::{io, net::SocketAddr};
+use std::{
+    future::Future,
+    io,
+    net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6},
+};
+
+use either::Either;
 
 pub async fn resolve_sock_addrs(
     host: &str,
@@ -68,4 +74,123 @@ fn to_addrs(mut result: *mut sys::addrinfo, port: u16) -> std::vec::IntoIter<Soc
         result = info.ai_next;
     }
     addrs.into_iter()
+}
+
+/// A trait for objects which can be converted or resolved to one or more
+/// [`SocketAddr`] values.
+///
+/// See [`std::net::ToSocketAddrs`].
+#[allow(async_fn_in_trait)]
+pub trait ToSocketAddrsAsync {
+    /// See [`std::net::ToSocketAddrs::Iter`].
+    type Iter: Iterator<Item = SocketAddr>;
+
+    /// See [`std::net::ToSocketAddrs::to_socket_addrs`].
+    async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter>;
+}
+
+// impl_to_sock_addrs_for_into_socket_addr
+macro_rules! itsafisa {
+    ($t:ty) => {
+        impl ToSocketAddrsAsync for $t {
+            type Iter = std::iter::Once<SocketAddr>;
+
+            async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter> {
+                Ok(std::iter::once(SocketAddr::from(*self)))
+            }
+        }
+    };
+}
+
+itsafisa!(SocketAddr);
+itsafisa!(SocketAddrV4);
+itsafisa!(SocketAddrV6);
+itsafisa!((IpAddr, u16));
+itsafisa!((Ipv4Addr, u16));
+itsafisa!((Ipv6Addr, u16));
+
+impl ToSocketAddrsAsync for (&str, u16) {
+    type Iter = Either<std::iter::Once<SocketAddr>, std::vec::IntoIter<SocketAddr>>;
+
+    async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter> {
+        let (host, port) = self;
+        if let Ok(addr) = host.parse::<Ipv4Addr>() {
+            return Ok(Either::Left(std::iter::once(SocketAddr::from((
+                addr, *port,
+            )))));
+        }
+        if let Ok(addr) = host.parse::<Ipv6Addr>() {
+            return Ok(Either::Left(std::iter::once(SocketAddr::from((
+                addr, *port,
+            )))));
+        }
+
+        resolve_sock_addrs(host, *port).await.map(Either::Right)
+    }
+}
+
+impl ToSocketAddrsAsync for (String, u16) {
+    type Iter = <(&'static str, u16) as ToSocketAddrsAsync>::Iter;
+
+    async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter> {
+        (&*self.0, self.1).to_socket_addrs_async().await
+    }
+}
+
+impl ToSocketAddrsAsync for str {
+    type Iter = <(&'static str, u16) as ToSocketAddrsAsync>::Iter;
+
+    async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter> {
+        if let Ok(addr) = self.parse::<SocketAddr>() {
+            return Ok(Either::Left(std::iter::once(addr)));
+        }
+
+        let (host, port_str) = self.rsplit_once(':').expect("invalid socket address");
+        let port: u16 = port_str.parse().expect("invalid port value");
+        (host, port).to_socket_addrs_async().await
+    }
+}
+
+impl ToSocketAddrsAsync for String {
+    type Iter = <(&'static str, u16) as ToSocketAddrsAsync>::Iter;
+
+    async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter> {
+        self.as_str().to_socket_addrs_async().await
+    }
+}
+
+impl<'a> ToSocketAddrsAsync for &'a [SocketAddr] {
+    type Iter = std::iter::Copied<std::slice::Iter<'a, SocketAddr>>;
+
+    async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter> {
+        Ok(self.iter().copied())
+    }
+}
+
+impl<T: ToSocketAddrsAsync + ?Sized> ToSocketAddrsAsync for &T {
+    type Iter = T::Iter;
+
+    async fn to_socket_addrs_async(&self) -> io::Result<Self::Iter> {
+        (**self).to_socket_addrs_async().await
+    }
+}
+
+pub async fn each_addr<T, F: Future<Output = io::Result<T>>>(
+    addr: impl ToSocketAddrsAsync,
+    mut f: impl FnMut(SocketAddr) -> F,
+) -> io::Result<T> {
+    let addrs = addr.to_socket_addrs_async().await?;
+    let mut last_err = None;
+    for addr in addrs {
+        match f(addr).await {
+            Ok(l) => return Ok(l),
+            Err(e) => last_err = Some(e),
+        }
+    }
+    Err(last_err.unwrap_or_else(|| {
+        io::Error::new(
+            io::ErrorKind::InvalidInput,
+            "could not resolve to any addresses",
+        )
+    }))
 }
